[create](router)Move functions to middleware{dvg}
