+++ .gitignore
@@ -1,3 +1,5 @@
+.DS_Store
+
 node_modules
 client/index.js
 client/index.js.map
