+++ server/routes/yelpRouter.js
@@ -1,0 +1,7 @@
+var router = require('express').Router();
+var yelpHandler = require('../handlers/yelpHandler.js');
+
+router.get('/locations', yelpHandler.getNearbyLocations);
+
+
+module.exports = router;
